refactor : cookie 관심사 분리
