+++ src/main/java/runnershigh/capstone/jwt/controller/JwtController.java
@@ -2,8 +2,8 @@ package runnershigh.capstone.jwt.controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import java.util.Objects;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.ResponseCookie;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -16,6 +16,7 @@ import runnershigh.capstone.jwt.dto.LoginRequest;
 import runnershigh.capstone.jwt.dto.LoginResponse;
 import runnershigh.capstone.jwt.enums.AuthConstants;
 import runnershigh.capstone.jwt.service.JwtService;
+import runnershigh.capstone.jwt.util.CookieUtil;
 
 @RestController
 @RequiredArgsConstructor
@@ -24,8 +25,7 @@ public class JwtController {
 
     private final JwtService jwtService;
     private final JwtProperties jwtProperties;
-
-    private static final String COOKIE_HEADER = "Set-Cookie";
+    private final CookieUtil cookieUtil;
 
     @PostMapping("/login")
     public LoginResponse login(@RequestBody LoginRequest loginRequest,
@@ -36,7 +36,7 @@ public class JwtController {
         response.setHeader(AuthConstants.AUTHORIZATION_HEADER.getValue(),
             AuthConstants.BEARER_PREFIX.getValue() + loginResponse.accessToken());
 
-        setRefreshTokenInCookie(response, loginResponse.refreshToken(),
+        cookieUtil.setRefreshTokenCookie(response, loginResponse.refreshToken(),
             jwtProperties.getRefreshExpirationTime());
 
         return loginResponse;
@@ -45,47 +45,25 @@ public class JwtController {
     @GetMapping("/user")
     public ResponseEntity<?> getUserFromToken(HttpServletRequest request) {
         String userId = (String) request.getAttribute("userId");
-        if (userId != null) {
-            return ResponseEntity.ok("Authenticated userId : " + userId);
-        } else {
+
+        if (Objects.isNull(userId)) {
             return ResponseEntity.status(401).body("Invalid or expired token");
         }
+        return ResponseEntity.ok("Authenticated userId : " + userId);
     }
 
     @DeleteMapping("/logout")
     public ResponseEntity<?> logout(HttpServletRequest request, HttpServletResponse response) {
         String userId = (String) request.getAttribute("userId");
 
-        if (userId == null) {
+        if (Objects.isNull(userId)) {
             return ResponseEntity.status(401).body("Invalid or expired token");
         }
 
         jwtService.logout(userId);
 
-        ResponseCookie expiredCookie = ResponseCookie.from(
-                AuthConstants.REFRESH_COOKIE_NAME.getValue(), "")
-            .maxAge(0)  // 쿠키 삭제
-            .path("/")
-            .secure(true)
-            .sameSite("None")
-            .httpOnly(true)
-            .build();
-
-        response.setHeader(COOKIE_HEADER, expiredCookie.toString());
+        cookieUtil.clearRefreshTokenCookie(response);
 
         return ResponseEntity.ok("Logged out successfully");
-    }
-
-    private void setRefreshTokenInCookie(HttpServletResponse response, String refreshToken,
-        Long refreshExpirationTime) {
-        ResponseCookie cookie = ResponseCookie.from(AuthConstants.REFRESH_COOKIE_NAME.getValue(),
-                refreshToken)
-            .maxAge(refreshExpirationTime)
-            .path("/")
-            .secure(true)
-            .sameSite("None")
-            .httpOnly(true)
-            .build();
-        response.setHeader(COOKIE_HEADER, cookie.toString());
     }
 }
