+++ src/main/java/runnershigh/capstone/jwt/enums/AuthConstants.java
@@ -4,7 +4,6 @@ public enum AuthConstants {
     REFRESH_COOKIE_NAME("refresh_token"),
     AUTHORIZATION_HEADER("Authorization"),
     BEARER_PREFIX("Bearer ");
-
     private final String value;
 
     AuthConstants(String value) {
